Move utxodag mock to the separate file
